Rename Footer to MainFooter
