+++ src/App.tsx
@@ -1,4 +1,4 @@
-import Footer from "@components/footer/Footer";
+import MainFooter from "@components/footer/MainFooter";
 import Navigation from "@components/navigation/Navigation";
 import { Fragment } from "react";
 import { Route, Routes } from "react-router-dom";
@@ -18,7 +18,7 @@ const App = () => {
         {/* TODO: create NoMatchPage component */}
         <Route path="*" element={<div>404 </div>} />
       </Routes>
-      <Footer />
+      <MainFooter />
     </Fragment>
   );
 };
